// source/frontend/src/models/api/generated/ApiGen_Concepts_CompReqLeaseProperty.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseAudit } from './ApiGen_Base_BaseAudit';
import { ApiGen_Concepts_PropertyLease } from './ApiGen_Concepts_PropertyLease';

// LINK: @backend/apimodels/Models/Concepts/CompensationRequisition/CompReqLeasePropertyModel.cs
export interface ApiGen_Concepts_CompReqLeaseProperty extends ApiGen_Base_BaseAudit {
  compensationRequisitionPropertyId: number | null;
  compensationRequisitionId: number | null;
  propertyLeaseId: number;
  leaseProperty: ApiGen_Concepts_PropertyLease | null;
}